[level 1] Title: 부족한 금액 계산하기, Time: 0.07 ms, Memory: 73.9 MB -BaekjoonHub

+++ 프로그래머스/lv1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -3,7 +3,7 @@ class Solution {
         
         long sum = 0;
         for(int i=1;i<=count;i++){
-            sum+=(long)price*i;
+            sum+=price*i;
         }
         return (money>=sum)? 0 : sum-money;
     }
